Move MappingError exceptions to base.exceptions
This avoids an upcoming import cycle
