+++ src/python/pants/engine/legacy/address_mapper.py
@@ -5,11 +5,11 @@ import logging
 import os
 
 from pants.base.build_file import BuildFile
+from pants.base.exceptions import ResolveError
 from pants.base.specs import AddressSpecs, DescendantAddresses, SiblingAddresses
 from pants.build_graph.address_lookup_error import AddressLookupError
 from pants.build_graph.address_mapper import AddressMapper
 from pants.engine.addressable import BuildFileAddresses
-from pants.engine.mapper import ResolveError
 from pants.util.dirutil import fast_relpath
