+++ tests/python/pants_test/engine/test_build_files.py
@@ -6,6 +6,7 @@ import re
 import unittest
 from typing import Tuple, Type, cast
 
+from pants.base.exceptions import ResolveError
 from pants.base.project_tree import Dir
 from pants.base.specs import AddressSpecs, SiblingAddresses, SingleAddress
 from pants.build_graph.address import Address
@@ -19,7 +20,7 @@ from pants.engine.build_files import (
 )
 from pants.engine.fs import Digest, FileContent, FilesContent, PathGlobs, Snapshot, create_fs_rules
 from pants.engine.legacy.structs import TargetAdaptor
-from pants.engine.mapper import AddressFamily, AddressMapper, ResolveError
+from pants.engine.mapper import AddressFamily, AddressMapper
 from pants.engine.nodes import Return, State, Throw
 from pants.engine.parser import HydratedStruct, SymbolTable
 from pants.engine.rules import rule
