+++ tests/python/pants_test/engine/test_mapper.py
@@ -6,19 +6,13 @@ import unittest
 from contextlib import contextmanager
 from textwrap import dedent
 
+from pants.base.exceptions import DuplicateNameError, UnaddressableObjectError
 from pants.base.specs import AddressSpec, AddressSpecs, SingleAddress
 from pants.build_graph.address import Address
 from pants.engine.addressable import BuildFileAddresses
 from pants.engine.build_files import create_graph_rules
 from pants.engine.fs import create_fs_rules
-from pants.engine.mapper import (
-  AddressFamily,
-  AddressMap,
-  AddressMapper,
-  DifferingFamiliesError,
-  DuplicateNameError,
-  UnaddressableObjectError,
-)
+from pants.engine.mapper import AddressFamily, AddressMap, AddressMapper, DifferingFamiliesError
 from pants.engine.objects import Collection
 from pants.engine.parser import HydratedStruct, SymbolTable
 from pants.engine.rules import rule
